pull prune_list out into standalone file
